⚡️ Common Adjustments

+++ src/app/pages/Submissions/index.js
@@ -11,17 +11,25 @@ import {
   CustomEditIconButton,
   CustomDeleteIconButton,
   ApproveIconButton,
+  CustomDetailButton,
 } from "app/components/CustomIconButton";
 import { DataGrid } from "@mui/x-data-grid";
 import { useNavigate } from "react-router-dom";
 import useFetch from "app/hooks/useFetch";
 import useAxiosFunction from "app/hooks/useAxiosFunction";
 import usePrevious from "app/hooks/usePrevious";
-import {isEqual} from "lodash";
+import { isEqual } from "lodash";
 import moment from "moment";
 import { withSnackbar } from "app/components/SnackbarComponent";
+import { withRoles } from "app/components/withRoles";
 
 const Submissions = (props) => {
+  const {
+    isAdminDireksi,
+    isAdminOperasional,
+    isAdminPerencanaan,
+    isSuperAdmin,
+  } = props;
   const { isLoading, data, error, axiosFetch } = useAxiosFunction();
   const {
     isLoading: isLoadingList,
@@ -77,7 +85,8 @@ const Submissions = (props) => {
         field: "period",
         headerName: "Periode",
         width: 150,
-        valueFormatter: (params) => params?.value ? moment(params?.value).format('DD MMMM YYYY') : "-",
+        valueFormatter: (params) =>
+          params?.value ? moment(params?.value).format("DD MMMM YYYY") : "-",
         sortable: false,
       },
       {
@@ -105,27 +114,49 @@ const Submissions = (props) => {
         headerName: "",
         type: "actions",
         getActions: (params) => {
-          return [
-            <ApproveIconButton
-              size="small"
-              sx={{ mr: 1 }}
-              onClick={() => approveData(params.row.id)}
-            />,
-            <CustomEditIconButton
-              size="small"
-              sx={{ mr: 1 }}
-              onClick={() => navigate(`/submissions/${params.row.id}/edit`)}
-            />,
-            <CustomDeleteIconButton
-              size="small"
-              onClick={() => deleteData(params.row.id)}
-            />,
-          ];
+          let actions = [];
+          if (isAdminDireksi) {
+            actions = [
+              <CustomDetailButton
+                size="small"
+                onClick={() => navigate(`/submissions/${params.row.id}/detail`)}
+              />,
+            ];
+          } else if (isAdminOperasional || isAdminPerencanaan || isSuperAdmin) {
+            actions = [
+              <CustomEditIconButton
+                size="small"
+                sx={{ mr: 1 }}
+                onClick={() => navigate(`/submissions/${params.row.id}/edit`)}
+              />,
+              <CustomDeleteIconButton
+                size="small"
+                onClick={() => deleteData(params.row.id)}
+              />,
+            ];
+            if (isAdminOperasional || isSuperAdmin) {
+              actions.unshift(
+                <ApproveIconButton
+                  size="small"
+                  sx={{ mr: 1 }}
+                  onClick={() => approveData(params.row.id)}
+                />
+              );
+            }
+          }
+          return actions;
         },
         width: 175,
       },
     ];
-  }, [currentPage, rowsPerPage]);
+  }, [
+    currentPage,
+    rowsPerPage,
+    isAdminDireksi,
+    isAdminOperasional,
+    isAdminPerencanaan,
+    isSuperAdmin,
+  ]);
 
   useEffect(() => {
     if (submissionData?.submission && submissionData?.paginator) {
@@ -161,7 +192,7 @@ const Submissions = (props) => {
         setFetched(true);
       },
     });
-  }
+  };
 
   const deleteData = (id) => {
     axiosFetch({
@@ -177,7 +208,7 @@ const Submissions = (props) => {
     });
   };
 
-  const onChangePage = useCallback((event, page) => {
+  const onChangePage = useCallback((page) => {
     setCurrentPage(page);
     setRequestParam((curr) => ({
       ...curr,
@@ -200,13 +231,15 @@ const Submissions = (props) => {
   return (
     <Fragment>
       <Box display="flex" justifyContent="flex-end" mb={4}>
-        <Button
-          type="button"
-          variant="contained"
-          onClick={() => navigate(`/submissions/new`)}
-        >
-          Tambah Data
-        </Button>
+        {isSuperAdmin || isAdminPerencanaan ? (
+          <Button
+            type="button"
+            variant="contained"
+            onClick={() => navigate(`/submissions/new`)}
+          >
+            Tambah Data
+          </Button>
+        ) : null}
       </Box>
       <DataGrid
         ref={tableRef}
@@ -223,7 +256,7 @@ const Submissions = (props) => {
         page={currentPage}
         pageSize={rowsPerPage}
         loading={isLoadingList}
-        // paginationMode="server"
+        paginationMode="server"
         rowCount={pagination?.itemCount ?? 10}
         rowsPerPageOptions={[10, 25, 50]}
       />
@@ -231,4 +264,4 @@ const Submissions = (props) => {
   );
 };
 
-export default withSnackbar(Submissions);
+export default withRoles(withSnackbar(Submissions));
